backend-repo/deepguard-backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
opencv-python-headless
requests
python-dotenv
httpx
google-generativeai
huggingface_hub
